Potential fix for code scanning alert no. 33: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type (size_t) to avoid overflow. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done using the larger type, and the result will be correctly assigned to the size_t variable.

The best way to fix this without changing existing functionality is to cast the first operand (dec->mt_method_ == 2 ? 2 : 1) to size_t before the multiplication. This change should be made on line 695 of the file libwebp-1.0.3/src/dec/frame_dec.c.

+++ libwebp-1.0.3/src/dec/frame_dec.c
@@ -692,7 +692,7 @@ static int AllocateMemory(VP8Decoder* const dec) {
         : 0;
   const size_t yuv_size = YUV_SIZE * sizeof(*dec->yuv_b_);
   const size_t mb_data_size =
-      (dec->mt_method_ == 2 ? 2 : 1) * mb_w * sizeof(*dec->mb_data_);
+      (size_t)(dec->mt_method_ == 2 ? 2 : 1) * mb_w * sizeof(*dec->mb_data_);
   const size_t cache_height = (16 * num_caches
                             + kFilterExtraRows[dec->filter_type_]) * 3 / 2;
   const size_t cache_size = top_size * cache_height;
